pyproject.toml
[project]
name = "rpg-guerrilha-urbana-backend"
version = "0.1.0"
description = "back pro rpg"
readme = "README.md"

authors = [
    {name = "Alice Campos", email = "allicefcampos@gmail.com"}
]

dependencies = [
    "obsidiantools>=0.23.0",
    "uvicorn[standard]>=0.22.0",
    "fastapi>=0.95.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "ruff>=0.0.241",
    "mypy>=1.3.0",
]


[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]

[tool.ruff]
line-length = 88
target-version = "py311"

exclude = [
    "venv",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]   

